+++ pymc3/distributions/multivariate.py
@@ -680,8 +680,8 @@ class DirichletMultinomial(Discrete):
 
     .. math::
 
-    f(x \mid n, a) = \frac{\Gamma(n + 1)\Gamma(\sum a_k)}
-                              {\Gamma(\n + \sum a_k)}
+        f(x \mid n, a) = \frac{\Gamma(n + 1)\Gamma(\sum a_k)}
+                              {\Gamma(n + \sum a_k)}
                          \prod_{k=1}^K
                          \frac{\Gamma(x_k +  a_k)}
                               {\Gamma(x_k + 1)\Gamma(a_k)}